Fixed for deletion of profile
- Fixed inter-service coms by adding missing environment  variables
- Fixed guest profile deletion
- Fixed wrong parsing of primitive.ObjectID

+++ reservation/handlers/reservationHandler.go
@@ -148,13 +148,18 @@ func (r *ReservationHandler) UpdateAvailablePeriodByAccommodation(rw http.Respon
 }
 
 func (r *ReservationHandler) DeletePeriodsForAccommodations(rw http.ResponseWriter, h *http.Request) {
-	accIDs := h.Context().Value(KeyProduct{}).([]primitive.ObjectID)
-	err := r.repo.DeletePeriodsForAccommodations(accIDs)
-	if err != nil {
-		r.logger.Print("Database exception: ", err)
-		rw.WriteHeader(http.StatusBadRequest)
-		return
+	if h.Context().Value(KeyProduct{}) != nil {
+		accIDs := h.Context().Value(KeyProduct{}).([]primitive.ObjectID)
+		if (accIDs != nil) && len(accIDs) > 0 {
+			err := r.repo.DeletePeriodsForAccommodations(accIDs)
+			if err != nil {
+				r.logger.Print("Database exception: ", err)
+				rw.WriteHeader(http.StatusBadRequest)
+				return
+			}
+		}
 	}
+
 	rw.WriteHeader(http.StatusNoContent)
 }
